requirements.txt

aioredis==1.3.0
asgiref==3.2.2
async-timeout==3.0.1
attrs==19.3.0
autobahn==19.10.1
Automat==0.7.0
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.13.0
channels==2.3.0
channels-redis==2.4.0
chardet==4.0.0
Click==7.0
constantly==15.1.0
cryptography==2.8
daphne==2.3.0
dash==1.9.1
dash-bootstrap-components==0.11.3
dash-core-components==1.8.1
dash-html-components==1.0.2
dash-renderer==1.2.4
dash-table==4.6.1
Django==2.2.6
django-bootstrap4==2.3.1
django-guardian==2.3.0
django-phonenumber-field==5.0.0
django-plotly-dash==1.0.1
django-redis==4.10.0
dpd-components==0.1.0
dpd-static-support==0.0.5
Flask==1.1.1
Flask-Compress==1.4.0
future==0.18.1
hiredis==1.0.0

idna==2.8
incremental==17.5.0
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
msgpack==0.6.2
numpy==1.19.5
pandas==1.1.5
phonenumbers==8.12.19
plotly==4.2.1
pycparser==2.19
PyHamcrest==1.9.0
python-dateutil==2.8.1
python-decouple==3.4
pytz==2019.3
redis==3.3.11
requests==2.25.1
retrying==1.3.3
six==1.12.0
soupsieve==2.2
sqlparse==0.3.0
Twisted==19.7.0
txaio==18.8.1
urllib3==1.26.3
Werkzeug==0.16.0
zope.interface==5.2.0
whitenoise==5.2.0
hyperlink==19.0.0
gunicorn==19.9.0